Moved TaigaClient out of taigapy/__init__.py
