+++ taigapy/consts.py
@@ -1,0 +1,6 @@
+DEFAULT_TAIGA_URL = "https://cds.team/taiga"
+
+# global variable to allow people to globally override the location before initializing client
+# which is often useful in adhoc scripts being submitted onto the cluster.
+DEFAULT_CACHE_DIR = "~/.taiga"
+CACHE_FILE = ".cache.db"
